Move browser start and close from setUp to Session
